refactor: add BlockServiceInterface to improve testability
